clean moss codec loader structure
