+++ sglang_omni/models/moss_tts/stages.py
@@ -3,15 +3,17 @@
 
 from __future__ import annotations
 
-import json
 import logging
-import os
-from contextlib import contextmanager
 from typing import Any
 
 import torch
 
 from sglang_omni.models.moss_tts.codec import split_moss_audio_segments
+from sglang_omni.models.moss_tts.hf_loading import (
+    load_moss_processor_class,
+    moss_transformers_processor_compat,
+    resolve_moss_checkpoint,
+)
 from sglang_omni.models.moss_tts.payload_types import (
     MossTTSState,
     moss_tts_special_token_defaults,
@@ -49,90 +51,8 @@ def store_state(payload: StagePayload, state: MossTTSState) -> StagePayload:
     return payload
 
 
-def _resolve_checkpoint(checkpoint: str) -> str:
-    if os.path.isdir(checkpoint):
-        return checkpoint
-    from huggingface_hub import snapshot_download
-
-    return snapshot_download(checkpoint)
-
-
 def _torch_dtype(dtype: str | torch.dtype) -> torch.dtype:
     return getattr(torch, dtype) if isinstance(dtype, str) else dtype
-
-
-@contextmanager
-def _moss_transformers_processor_compat():
-    """Scope load-time Transformers API-drift patches to the MOSS processor load
-    (renamed PreTrainedConfig, the audio-tokenizer modality mapping, and
-    PreTrainedAudioTokenizerBase), restoring the globals afterwards so they don't
-    leak into unrelated models."""
-    import transformers.configuration_utils as configuration_utils
-    from transformers import PreTrainedModel, processing_utils
-
-    missing = object()
-    undo: list[tuple[str, Any, str, Any]] = []
-
-    def patch_attr(obj: Any, name: str, value: Any) -> None:
-        undo.append(("attr", obj, name, getattr(obj, name, missing)))
-        setattr(obj, name, value)
-
-    def patch_item(mapping: dict, key: str, value: Any) -> None:
-        undo.append(("item", mapping, key, mapping.get(key, missing)))
-        mapping[key] = value
-
-    try:
-        if not hasattr(configuration_utils, "PreTrainedConfig"):
-            patch_attr(
-                configuration_utils,
-                "PreTrainedConfig",
-                configuration_utils.PretrainedConfig,
-            )
-        auto_mapping = getattr(processing_utils, "AUTO_TO_BASE_CLASS_MAPPING", None)
-        if isinstance(auto_mapping, dict):
-            if "AutoModel" not in auto_mapping:
-                patch_item(auto_mapping, "AutoModel", "PreTrainedModel")
-            if not hasattr(processing_utils, "MODALITY_TO_BASE_CLASS_MAPPING"):
-                patch_attr(
-                    processing_utils, "MODALITY_TO_BASE_CLASS_MAPPING", auto_mapping
-                )
-        if hasattr(processing_utils, "PreTrainedAudioTokenizerBase"):
-            patch_attr(
-                processing_utils, "PreTrainedAudioTokenizerBase", PreTrainedModel
-            )
-        yield
-    finally:
-        for kind, obj, key, old in reversed(undo):
-            if kind == "attr":
-                if old is missing:
-                    if hasattr(obj, key):
-                        delattr(obj, key)
-                else:
-                    setattr(obj, key, old)
-            elif old is missing:
-                obj.pop(key, None)
-            else:
-                obj[key] = old
-
-
-def _load_moss_processor_class(checkpoint_dir: str) -> type:
-    from transformers.dynamic_module_utils import get_class_from_dynamic_module
-
-    processor_config_path = os.path.join(checkpoint_dir, "processor_config.json")
-    with open(processor_config_path, encoding="utf-8") as f:
-        processor_config = json.load(f)
-
-    class_ref = (processor_config.get("auto_map") or {}).get("AutoProcessor")
-    if not class_ref:
-        raise RuntimeError("MOSS-TTS processor_config.json lacks AutoProcessor map")
-
-    processor_cls = get_class_from_dynamic_module(class_ref, checkpoint_dir)
-    if list(getattr(processor_cls, "attributes", [])) == [
-        "feature_extractor",
-        "tokenizer",
-    ]:
-        processor_cls.attributes = ["tokenizer"]
-    return processor_cls
 
 
 def _normalize_moss_processor_config(processor: Any) -> None:
@@ -151,11 +71,11 @@ def _load_moss_processor(
     device: str = "cpu",
     dtype: str | torch.dtype = "float32",
 ) -> Any:
-    checkpoint_dir = _resolve_checkpoint(model_path)
+    checkpoint_dir = resolve_moss_checkpoint(model_path)
     logger.info("Loading MOSS-TTS processor from %s on %s", checkpoint_dir, device)
     try:
-        with _moss_transformers_processor_compat():
-            processor_cls = _load_moss_processor_class(checkpoint_dir)
+        with moss_transformers_processor_compat():
+            processor_cls = load_moss_processor_class(checkpoint_dir)
             processor = processor_cls.from_pretrained(
                 checkpoint_dir,
                 trust_remote_code=True,
@@ -224,7 +144,7 @@ def create_sglang_tts_engine_executor(
         build_sglang_server_args,
     )
 
-    checkpoint_dir = _resolve_checkpoint(model_path)
+    checkpoint_dir = resolve_moss_checkpoint(model_path)
     if gpu_id is not None:
         device = f"cuda:{gpu_id}"
     gpu_id = int(device.split(":")[-1]) if ":" in device else 0
